+++ History.md
@@ -1,3 +1,8 @@
+### 0.2.1 / 2013-08-30
+
+* included resources can be functions or objects with `execute` method
+* added ability to customize inclusion by using `data-` attributes and functional inclusions
+
 ### 0.2.0 / 2013-08-21
 
 * added support for `x-include` and `x-require` dependency directives